+++ src/test/java/janggi/support/TestCollisionDetector.java
@@ -1,9 +1,10 @@
-package janggi.domain.rule.collision;
+package janggi.support;
 
 import janggi.domain.piece.Piece;
+import janggi.domain.rule.collision.CollisionDetector;
 import java.util.List;
 
-public class TestCollisionDetector implements CollisionDetector{
+public class TestCollisionDetector implements CollisionDetector {
     @Override
     public void check(Piece piece, List<Piece> piecesOnPath) {
         throw new UnsupportedOperationException();
